Data for Germany, a given state and the r-value can now be obtained independent of each other. Split `getData()` into `getRData()`, `getDataGer()`, `getDataState()` and `getDataAreaState()`. ( `get<Name>Data()` will return pure data, `getData<Name>()` will return a `DataResponse` object.)
Values for rValue, country and states will be cached for the runtime of the script. For example multiple request to states data will result in only one request to the api.

Added new function `prepend()` to cast a given value to a string an prepend it n-times with the given value.